requirements.txt
# Streamlit dependencies
streamlit>=1.37
pandas>=2.1
numpy>=1.26
pydantic>=2.6
python-dotenv>=1.0
mistralai>=1.0
openpyxl>=3.1
rank-bm25>=0.2.2
pyyaml>=6.0

# Widget service dependencies
fastapi==0.115.6
uvicorn[standard]==0.32.1
httpx==0.27.2
jinja2==3.1.4
